requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
influxdb-client==1.40.0
python-dateutil==2.8.2
